Fix Unmarshal Response Issue
Unfortunately defined type int64 for a variable which is actually float64
Struct and response example got updated

+++ api_integrations/outdoor_temperature_viessmann/struct.go
@@ -15,9 +15,9 @@ type OutsideTemperature struct {
 				Value string `json:"value"`
 			} `json:"status"`
 			Value struct {
-				Type  string `json:"type"`
-				Unit  string `json:"unit"`
-				Value int64  `json:"value"`
+				Type  string  `json:"type"`
+				Unit  string  `json:"unit"`
+				Value float64 `json:"value"`
 			} `json:"value"`
 		} `json:"properties"`
 		Timestamp string `json:"timestamp"`
